Move ContextProvider with other components
